refactor: replace teamStatusWindow logic to separate file
